feat(web,sdf) move proposed actions to new endpoint
