+++ app/web/src/api/sdf/dal/change_set.ts
@@ -1,4 +1,3 @@
-import { ProposedAction, ActionId } from "@/store/actions.store";
 import { UserId } from "@/store/auth.store";
 
 export enum ChangeSetStatus {
@@ -15,7 +14,6 @@ export interface ChangeSet {
   id: ChangeSetId;
   pk: ChangeSetId;
   name: string;
-  actions: Record<ActionId, ProposedAction>;
   status: ChangeSetStatus;
   appliedByUserId?: UserId;
   appliedAt?: IsoDateString;
